+++ changes.md
@@ -1,3 +1,12 @@
+# V3.2.0
+## Added
+- The option to use different languages (There is Dutch and English now).
+- The option to use a other message file in stead of the default one.
+- The client.stop() function.
+## Fixed
+- The colors for the permission error message embed.
+- The bot permissions message.
+
 # V3.1.4
 ## Added
 - The changes log file